+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 _See also: [the official libGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
-#### 1.10.0-SNAPSHOT
+#### 1.10.0-rc2
 
 - **[UPDATE]** Updated to Kotlin 1.6.21.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.6.1.